refactor : ContentType Enum domain 패키지에서 constant 로 이동
