+++ src/main/java/qna/domain/DeleteHistory.java
@@ -14,6 +14,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import qna.constant.ContentType;
 import qna.domain.collections.Answers;
 import qna.domain.collections.DeleteHistories;
