+++ src/test/java/qna/domain/repository/DeleteHistoryRepositoryTest.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import qna.domain.ContentType;
+import qna.constant.ContentType;
 import qna.domain.DeleteHistory;
 import qna.domain.User;
