+++ src/test/java/qna/service/QnaServiceTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import qna.constant.ContentType;
 import qna.exception.CannotDeleteException;
 import qna.domain.*;
